requirements.txt
mysql-connector-python
tiktoken 
python-dotenv
openai
fastapi 
uvicorn